+++ services-directions-models/src/test/java/com/mapbox/api/directions/v5/models/DirectionsRouteTest.java
@@ -9,6 +9,9 @@ import com.mapbox.geojson.Point;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
 import org.hamcrest.junit.ExpectedException;
 import org.junit.Rule;
 import org.junit.Test;
@@ -130,9 +133,7 @@ public class DirectionsRouteTest extends TestUtils {
     String uuid = "123";
     DirectionsRoute route = DirectionsRoute.fromJson(json, options, uuid);
 
-    DirectionsRoute newRoute = route.toBuilder().routeIndex("0").build();
-
-    assertEquals(2, newRoute.tollCosts().size());
+    assertEquals(2, route.tollCosts().size());
   }
 
   @Test
@@ -148,8 +149,72 @@ public class DirectionsRouteTest extends TestUtils {
     String uuid = "123";
     DirectionsRoute route = DirectionsRoute.fromJson(json, options, uuid);
 
-    DirectionsRoute newRoute = route.toBuilder().routeIndex("0").build();
+    assertNull(route.tollCosts());
+  }
 
-    assertNull(newRoute.tollCosts());
+  @Test
+  public void directionsRoute_hasWaypoints() throws IOException {
+    String json = loadJsonFixture("directions_v5_with_waypoints.json");
+    RouteOptions options = RouteOptions.builder()
+      .profile(DirectionsCriteria.PROFILE_DRIVING_TRAFFIC)
+      .coordinatesList(new ArrayList<Point>() {{
+        add(Point.fromLngLat(1.0, 1.0));
+        add(Point.fromLngLat(2.0, 2.0));
+      }})
+      .waypointsPerRoute(true)
+      .build();
+    String uuid = "123";
+    DirectionsRoute route = DirectionsRoute.fromJson(json, options, uuid);
+
+    assertEquals(2, route.waypoints().size());
+  }
+
+  @Test
+  public void directionsRouteBuilder_hasWaypoints() {
+    List<DirectionsWaypoint> waypoints = Arrays.asList(
+      DirectionsWaypoint.builder()
+        .name("name1")
+        .rawLocation(new double[] { 5.6, 7.8 })
+        .build(),
+      DirectionsWaypoint.builder()
+        .name("name2")
+        .rawLocation(new double[] { 1.2, 3.4 })
+        .build()
+    );
+    DirectionsRoute route = DirectionsRoute.builder()
+      .duration(12.12)
+      .distance(34.34)
+      .waypoints(waypoints)
+      .build();
+
+    assertEquals(waypoints, route.waypoints());
+  }
+
+
+  @Test
+  public void directionsRoute_noWaypoints() throws IOException {
+    String json = loadJsonFixture("directions_v5-with-closure_precision_6.json");
+    RouteOptions options = RouteOptions.builder()
+      .profile(DirectionsCriteria.PROFILE_DRIVING_TRAFFIC)
+      .coordinatesList(new ArrayList<Point>() {{
+        add(Point.fromLngLat(1.0, 1.0));
+        add(Point.fromLngLat(2.0, 2.0));
+      }})
+      .waypointsPerRoute(true)
+      .build();
+    String uuid = "123";
+    DirectionsRoute route = DirectionsRoute.fromJson(json, options, uuid);
+
+    assertNull(route.waypoints());
+  }
+
+  @Test
+  public void directionsRouteBuilder_noWaypoints() {
+    DirectionsRoute route = DirectionsRoute.builder()
+      .duration(12.12)
+      .distance(34.34)
+      .build();
+
+    assertNull(route.waypoints());
   }
 }